refactor(user-media): unify recos and bookmarks handling and improve cache management
- Renamed `WidgetUserWatchlist` to `WidgetBookmarks` and `WidgetUserRecos` to `WidgetRecos` for consistency.
- Consolidated media-specific reco mutations (e.g., `useUserRecosMovieDeleteMutation`) into generic `useUserRecoDeleteByMediaMutation` and `useUserRecoDeleteByIdMutation`.
- Enhanced cache invalidation for reco mutations:
    - `useUserRecoDeleteByMediaMutation` now removes reco items from all related caches.
    - `useUserRecoDeleteByIdMutation` selectively removes a sender from a reco or the reco entirely if it's the last sender, updating `latestCreatedAt`.
- Introduced `updateOrRemoveListItemInAllCaches` utility to enable flexible cache updates and removals.
- Updated UI and query client logic to use the new unified reco and bookmark hooks.
- Added explicit `::int` casting to the `COALESCE` SQL function in `UserRecosService` to ensure correct type handling.
- Improved accessibility in `ModalRecoSenders` by adding a screen-reader-only modal description.

This refactoring centralizes logic, improves type safety, and optimizes data consistency across the application by standardizing how user recommendations and bookmarks are managed and cached.

+++ apps/api/src/app/users/recos/user-recos.service.ts
@@ -27,7 +27,7 @@ export class UserRecosService {
     }
 
     return tx.select({
-        mediaId: sql<number>`COALESCE(${reco.movieId}, ${reco.tvSeriesId})`.as('media_id'),
+        mediaId: sql<number>`COALESCE(${reco.movieId}, ${reco.tvSeriesId})::int`.as('media_id'),
         type: reco.type,
         firstSendAt: sql<string>`MIN(${reco.createdAt})`.as('first_send_at'),
         lastSendAt: sql<string>`MAX(${reco.createdAt})`.as('last_send_at'),
